feat(avo-2597): Extract share with colleague content from share dropdown and create separate component, split scss by component, add navy color to variables
