+++ src/shared/components/ShareDropdown/ShareDropdown.scss
@@ -1,7 +1,6 @@
 @import '../../../styles/settings/variables';
 
 $border: 1px solid $color-gray-200;
-$borderLight: 1px solid $color-gray-100;
 
 .c-assignment-heading__actions .c-dropdown__content-open {
 	padding-top: 0 !important;
@@ -9,10 +8,11 @@ $borderLight: 1px solid $color-gray-100;
 
 .c-dropdown__content-open {
 	border: $border;
+	border-radius: 0.5rem;
 	background-color: $g-base-background;
 	margin-top: 1rem;
 	width: 545px;
-	// background-color: red;
+	filter: drop-shadow(0px 4px 4px rgba($color-navy, 0.25));
 
 	.c-tabs {
 		width: 100%;
@@ -32,6 +32,11 @@ $borderLight: 1px solid $color-gray-100;
 
 			&:first-child {
 				border-right: $border;
+				border-radius: 0.5rem 0 0 0;
+			}
+
+			&:last-child {
+				border-radius: 0 0.5rem 0 0;
 			}
 
 			&__active {
@@ -45,101 +50,4 @@ $borderLight: 1px solid $color-gray-100;
 
 .c-share-dropdown__content {
 	padding: 2rem;
-}
-
-.c-add-colleague {
-	display: flex;
-
-	.c-button.c-button--primary {
-		border-radius: 0;
-		// flex-grow: 1;
-		color: $color-gray-1000;
-
-		svg * {
-			fill: $color-gray-1000;
-		}
-	}
-
-	&__select.c-button.c-button--primary {
-		width: 100%;
-		border: $borderLight;
-		border-right: none;
-		background-color: $g-base-background;
-		text-transform: capitalize;
-
-		&:hover,
-		:active,
-		:active:focus {
-			background-color: $g-base-background;
-			border-color: $color-gray-100;
-		}
-	}
-
-	.c-input {
-		flex-grow: 1;
-		border: $borderLight;
-		border-right: none;
-		border-radius: $c-input-border-radius 0 0 $c-input-border-radius;
-	}
-
-	.c-dropdown__trigger {
-		min-width: max-content;
-	}
-
-	&__button.c-button.c-button--primary {
-		min-width: max-content;
-		border: $borderLight;
-		border-radius: 0 $c-input-border-radius $c-input-border-radius 0;
-		background-color: $color-gray-50;
-	}
-
-	.c-dropdown__content-open {
-		border: $borderLight;
-		margin-top: 0;
-		width: fit-content;
-	}
-}
-
-.c-colleagues-info-list {
-	margin-top: 2rem;
-	display: grid;
-	row-gap: 1rem;
-}
-
-.c-colleague-info-row {
-	display: grid;
-	align-items: center;
-	column-gap: 1rem;
-	grid-template-columns: max-content 1fr 1fr 7%;
-	font-size: 15px;
-
-	&__avatar {
-		.c-avatar img {
-			width: 100%;
-			height: 100%;
-		}
-	}
-
-	&__info {
-		&__email {
-			font-size: 14px;
-			color: $color-gray-200;
-		}
-	}
-
-	&__rights {
-		justify-self: flex-end;
-		text-transform: capitalize;
-	}
-
-	&__action {
-		justify-self: flex-end;
-	}
-}
-
-.c-icon-button {
-	margin-left: 0.5rem;
-	background-color: $g-base-background;
-	border: none;
-	cursor: pointer;
 }
