+++ src/shared/components/ShareWithColleagues/ShareWithColleagues.scss
@@ -1,0 +1,109 @@
+@import '../../../styles/settings/variables';
+
+$borderLight: 1px solid $color-gray-100;
+
+.c-add-colleague {
+	display: flex;
+
+	.c-button.c-button--primary {
+		border-radius: 0;
+		color: $color-navy;
+
+		svg * {
+			fill: $color-navy;
+		}
+	}
+
+	&__error {
+		margin-top: 0.5rem;
+		color: $color-error;
+	}
+
+	&__select.c-button.c-button--primary {
+		width: 100%;
+		border: $borderLight;
+		border-right: none;
+		background-color: $g-base-background;
+		text-transform: capitalize;
+
+		&:hover,
+		:active,
+		:active:focus {
+			background-color: $g-base-background;
+			border-color: $color-gray-100;
+		}
+	}
+
+	.c-input {
+		flex-grow: 1;
+		border: $borderLight;
+		border-right: none;
+		border-radius: $c-input-border-radius 0 0 $c-input-border-radius;
+	}
+
+	.c-dropdown__trigger {
+		min-width: max-content;
+	}
+
+	&__button.c-button.c-button--primary {
+		min-width: max-content;
+		border: $borderLight;
+		border-radius: 0 $c-input-border-radius $c-input-border-radius 0;
+		background-color: $color-gray-50;
+	}
+
+	.c-dropdown__content-open {
+		border: $borderLight;
+		margin-top: 0;
+		width: fit-content;
+	}
+}
+
+.c-colleagues-info-list {
+	margin-top: 2rem;
+	display: grid;
+	row-gap: 1rem;
+}
+
+.c-colleague-info-row {
+	display: grid;
+	align-items: center;
+	column-gap: 1rem;
+	grid-template-columns: max-content 1fr 1fr 7%;
+	font-size: 15px;
+	color: $color-navy;
+
+	&__avatar {
+		.c-avatar img {
+			width: 100%;
+			height: 100%;
+		}
+	}
+
+	&__info {
+		&__email {
+			font-size: 14px;
+			color: $color-gray-200;
+		}
+	}
+
+	&__rights {
+		justify-self: flex-end;
+		text-transform: capitalize;
+	}
+
+	&__action {
+		justify-self: flex-end;
+	}
+}
+
+.c-icon-button {
+	margin-left: 0.5rem;
+	background-color: $g-base-background;
+	border: none;
+	cursor: pointer;
+
+	svg {
+		fill: $color-navy;
+	}
+}
