+++ src/shared/components/index.ts
@@ -21,3 +21,4 @@ export * from './CheckboxDropdownModal/CheckboxDropdownModal';
 export * from './ListSorter';
 export * from './LoadingErrorLoadedComponent/LoadingErrorLoadedComponent';
 export { default as ShareDropdown } from './ShareDropdown/ShareDropdown';
+export { default as ShareWithColleagues } from './ShareWithColleagues/ShareWithColleagues';
