+++ src/shared/translations/nl.json
@@ -1661,6 +1661,9 @@
   "shared/components/quick-lane-modal/quick-lane-modal___snel-delen": "Delen met leerlingen",
   "shared/components/quick-lane-modal/quick-lane-modal___titel": "Titel",
   "shared/components/quick-lane-modal/quick-lane-modal___weergave-voor-leerlingen": "Weergave voor leerlingen",
+  "shared/components/share-dropdown/share-dropdown___collegas": "Collega's",
+  "shared/components/share-dropdown/share-dropdown___delen": "Delen",
+  "shared/components/share-dropdown/share-dropdown___leerlingen": "Leerlingen",
   "shared/components/share-through-email-modal/share-through-email-modal___de-email-is-verstuurd": "De e-mail is verstuurd",
   "shared/components/share-through-email-modal/share-through-email-modal___de-url-is-naar-het-klembord-gekopieerd": "De hyperlink is naar het klembord gekopieerd.",
   "shared/components/share-through-email-modal/share-through-email-modal___kopieer-deze-publieke-link": "Kopieer deze publieke link",
@@ -1671,6 +1674,12 @@
   "shared/components/share-through-email-modal/share-through-email-modal___verzenden": "Verzenden",
   "shared/components/share-through-email-modal/share-through-email-modal___waarschuwing-deel-link-via-email": "Opgelet, enkel gebruikers met een account op Het Archief voor Onderwijs zullen deze link kunnen openen.",
   "shared/components/share-through-email-modal/share-through-email-modal___wij-sturen-voor-jou-een-mailtje-met-deze-link": "Wij sturen voor jou een mail met deze link.",
+  "shared/components/share-with-colleagues/share-with-colleagues___bijdrager": "Bijdrager",
+  "shared/components/share-with-colleagues/share-with-colleagues___email-is-geen-geldig-emailadres": "Email is geen geldig emailadres",
+  "shared/components/share-with-colleagues/share-with-colleagues___emailadres": "Emailadres",
+  "shared/components/share-with-colleagues/share-with-colleagues___kijker": "Kijker",
+  "shared/components/share-with-colleagues/share-with-colleagues___rol": "Rol",
+  "shared/components/share-with-colleagues/share-with-colleagues___voeg-toe": "Voeg toe",
   "shared/components/wysiwyg-wrapper/wysiwyg-wrapper___dit-bestand-is-te-groot-max-10-mb": "Dit bestand is te groot (max. 10MB).",
   "shared/components/wysiwyg-wrapper/wysiwyg-wrapper___het-opladen-van-de-afbeelding-is-mislukt": "Het opladen van de afbeelding is mislukt.",
   "shared/constants/index___meer-opties": "Meer opties",
