+++ src/styles/settings/_colors.scss
@@ -29,6 +29,7 @@ $color-dark-night-blue: #182f42;
 $color-teal-bright: #25a4cf;
 $color-teal: #1d637a;
 $color-teal-dark: #124455;
+$color-navy: #374a5c;
 
 /* Secondary colors
    ========================================================================== */
